requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
websockets>=12.0
requests==2.32.3
aiofiles==23.2.1
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
reportlab
python-dotenv
duckduckgo-search
python-multipart
pdfplumber
PyPDF2
python-docx
pillow
pytesseract
rarfile
mypy==1.10.0
types-requests==2.32.0.20241016

# LLM client
openai>=1.30.0,<2

# Advanced Memory System
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
networkx>=3.0
numpy>=1.24.0
